pyproject.toml
[project]
name = "nisaa"
version = "0.1.0"
description = ""
authors = [
    {name = "priyanshu",email = "priyanshu.sahu@nzminds.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-dotenv>=1.0.0",
    "requests>=2.32.3,<3.0.0",
    "python-multipart (==0.0.20)",
    "langgraph (>=1.0.1,<2.0.0)",
    "openai (>=2.6.0,<3.0.0)",
    "numpy (>=2.3.4,<3.0.0)",
    "scikit-learn (>=1.7.2,<2.0.0)",
    "pinecone (>=7.3.0,<8.0.0)",
    "langchain (>=1.0.2,<2.0.0)",
    "langchain-openai (>=1.0.1,<2.0.0)",
    "langgraph-checkpoint-postgres (>=3.0.0,<4.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
]

[tool.poetry]
packages = [{include = "nisaa", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
